continued refactoring
